fix tests: wait for load (and move to WithHead)
